+++ Scripts/exec-login-edison.sh
@@ -8,8 +8,8 @@
 #IP_EDISON=192.168.205.220  #IP to network LCR
 #IP_EDISON=172.28.107.227   #IP to network eduroam
 #IP_EDISON=192.168.0.102    #IP to network RedeDrone
-IP_EDISON=192.168.43.133   #IP to network redeDroneC (celular veronica)
-#IP_EDISON=192.168.43.9     #IP to network RedeDroneCjes (celular jesimar)
+IP_EDISON=192.168.43.133   #IP to network redeDroneC (cell phone veronica)
+#IP_EDISON=192.168.43.9     #IP to network RedeDroneCjes (cell phone jesimar)
 #IP_EDISON=10.42.0.199      #IP to network RedeDronePC
 #IP_EDISON=192.168.0.100    #IP to network Alice