+++ Scripts/exec-sitl.sh
@@ -5,10 +5,19 @@
 #Description: Script that runs Dronekit Software-In-The-Loop (SITL) to do simulations.
 #Descrição: Script que executa o Dronekit Software-In-The-Loop (SITL) para fazer simulações.
 
-LAT=-22.00593264981567  #latitude inicial do drone, por exemplo: -22.005640
-LNG=-47.89870966454083  #longitude inicial do drone, por exemplo: -47.932474
-ALT=870					#altitude absoluta inicial do drone em metros, por exemplo: 870
-ANGLE=0					#ângulo (orientação) inicial do drone em graus, por exemplo: 0   (0 graus é norte)
-SPEED=1.0 				#velocidade do drone durante a missão simulada, por exemplo: 1.0 (1.0 é velocidade normal)
+#initial latitude of the drone, for example: -22.00593264981567, -22.005640
+LAT=-22.00597264465543
+
+#initial longitude of the drone, for example: -47.89870966454083, -47.932474
+LNG=-47.89868819614218
+
+#initial absolute altitude of the drone in meters, for example: 870
+ALT=870
+
+#angle (orientation) of the drone in degrees, for example: 0   (0 degrees is north)
+ANGLE=90
+
+#drone speed during the simulated mission, for example: 1.0 (1.0 is normal speed)
+SPEED=1.0
 
 dronekit-sitl copter-3.3 --home=$LAT,$LNG,$ALT,$ANGLE --speedup=$SPEED