+++ UAV-Generic/src/uav/generic/struct/constants/TypeOperationMode.java
@@ -1,0 +1,13 @@
+package uav.generic.struct.constants;
+
+/**
+ *
+ * @author Jesimar S. Arantes
+ */
+public class TypeOperationMode {
+   
+    public static final String SITL_LOCAL  = "SITL_LOCAL";
+    public static final String SITL_EDISON = "SITL_EDISON";
+    public static final String REAL_FLIGHT = "REAL_FLIGHT";
+    
+}