+++ UAV-Mission-Creator/src/uav/mission_creator/struct/geom/Point.java
@@ -1,0 +1,13 @@
+package uav.mission_creator.struct.geom;
+
+import java.io.Serializable;
+
+/**
+ *
+ * @author Jesimar S. Arantes
+ */
+public abstract class Point implements Serializable{
+    
+    public abstract double distance(Point point);
+    
+}